feat: 优化文件传输逻辑，增强错误处理，添加文件大小属性，调整日志输出级别

+++ app/tree/buildFileTree.go
@@ -36,7 +36,7 @@ func getLeafInfo(filepath string) *Node {
 		Name:     fileInfo.Name(),
 		ParentID: "",
 		IsDir:    fileInfo.IsDir(),
-		Size:     fileInfo.Size(),
+		Size:     uint64(fileInfo.Size()),
 		ModTime:  fileInfo.ModTime(),
 		Hash:     "",
 	}
